Add configResolver and enable stepping
ConfigResolver is used to take a launch.json config and turn it into a
config used to launch C-SPY. Will need much more work before this
actually works.

Stepping is simply done by sending step requests to cspyserver and
listening for kDkTargetStopped events so we know when to send
StoppedEvent s to the DAP client.