Rewrite registers code to support expandable regs
The ListWindowClient class now keeps track of the tree state of the list
window (i.e. what rows are expandable, what their children are).
This is then used to implement ListWindowVariablesProvider, which
converts list window contents into DAP Variables.
This also requires some management of variableReferences, to make sure
that all references sent to the client are created by the same Handles
instance (the one in CSpyContextManager.